+++ src/test/java/io/personium/core/bar/AllTests.java
@@ -21,7 +21,7 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 /**
- * パッケージ配下のテストケースを全て実行するためのテストスイート.
+ * Test suite for executing all test cases under this package.
  */
 @RunWith(Suite.class)
 @SuiteClasses({
