+++ src/test/java/io/personium/core/model/impl/es/AllTests.java
@@ -21,11 +21,12 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 /**
- * パッケージ配下のテストケースを全て実行するためのテストスイート.
+ * Test suite for executing all test cases under this package.
  */
 @RunWith(Suite.class)
 @SuiteClasses({
-    EsModelTest.class
+    EsModelTest.class,
+    CellEsImplTest.class
     })
 public class AllTests {
 }
