+++ src/test/java/io/personium/core/rs/AllTests.java
@@ -21,11 +21,12 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 /**
- * パッケージ配下のテストケースを全て実行するためのテストスイート.
+ * Test suite for executing all the test cases under this package.
  */
 @RunWith(Suite.class)
 @SuiteClasses({
-    PersoniumCoreExceptionMapperTest.class
+    PersoniumCoreExceptionMapperTest.class,
+    PersoniumCoreContainerFilterTest.class
     })
 public class AllTests {
 }
