+++ src/test/java/io/personium/test/unit/core/jersey/filter/AllTests.java
@@ -20,6 +20,8 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
+import io.personium.core.rs.PersoniumCoreContainerFilterTest;
+
 /**
  * パッケージ配下のテストケースを全て実行するためのテストスイート.
  */
